chore: add `auto_route` package
