+++ lib/pages/pages.dart
@@ -1,0 +1,1 @@
+export 'home_page.dart';
