+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   http: ^1.1.0
   uri: ^1.0.0
   wakelock_plus: ^1.1.1
+  auto_route: ^7.8.2
 
 dev_dependencies:
   flutter_test:
@@ -51,6 +52,7 @@ dev_dependencies:
   riverpod_lint: ^2.0.1
   freezed: ^2.4.2
   json_serializable: ^6.7.1
+  auto_route_generator: ^7.3.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
